exact visibility rules
